WIP character functionality

+++ src/components/DeleteConfirmPopup/DeleteConfirmPopup.tsx
@@ -2,6 +2,7 @@ import { useSelector } from 'react-redux';
 import crossImg from '../../assets/cross.svg';
 import type { RootState } from '../../store';
 import useAttack from '../../utils/hooks/useAttack';
+import useCharacter from '../../utils/hooks/useCharacter';
 import usePopup from '../../utils/hooks/usePopup';
 import styles from './DeleteConfirmPopup.module.css';
 
@@ -9,10 +10,11 @@ export default function DeleteConfirmPopup() {
 
     const { hideDeleteConfirmPopup } = usePopup()
     const { deleteAttack } = useAttack();
-    const { attacks } = useSelector((state: RootState) => state.data)
-    const { deleteAttackIndex } = useSelector((state: RootState) => state.popup)
+    const { deleteCharacter } = useCharacter();
+    const { attacks, characters } = useSelector((state: RootState) => state.data)
+    const { deleteAttackIndex, deleteCharacterIndex, deleteActionType } = useSelector((state: RootState) => state.popup)
 
-    const attackToBeDeleted = attacks[deleteAttackIndex ?? -1]
+    const itemToBeDeleted = deleteActionType === "attack" ? attacks[deleteAttackIndex ?? -1] : characters[deleteCharacterIndex ?? -1]
 
     const handleClose = () => {
         hideDeleteConfirmPopup()
@@ -23,8 +25,16 @@ export default function DeleteConfirmPopup() {
     }
 
     const handleYes = () => {
-        deleteAttack(deleteAttackIndex ?? -1);
         hideDeleteConfirmPopup()
+
+        if (deleteActionType === "attack") {
+            deleteAttack(deleteAttackIndex ?? -1);
+        }
+
+        if (deleteActionType === "character") {
+            deleteCharacter(deleteCharacterIndex ?? -1)
+        }
+
     }
 
 
@@ -32,7 +42,7 @@ export default function DeleteConfirmPopup() {
         <div className={styles.Overlay}>
             <div className={styles.DeleteConfirmPopup}>
                 <div className={styles.TitleContainer}>
-                    <p>Are you sure you want to delete {attackToBeDeleted.name}?</p>
+                    <p>Are you sure you want to delete {itemToBeDeleted.name}?</p>
                     <img src={crossImg} alt="Close" onClick={handleClose} />
                 </div>
                 <div className={styles.ButtonContainer}>
